Move SearchInput to components/
